+++ tests/test_monergism_scrap_metadata.py
@@ -1,3 +1,4 @@
+import asyncio
 import os 
 import sys 
 import pathlib
@@ -5,6 +6,7 @@ import pathlib
 import unittest
 
 from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_metadata
+from Instrumentum_sanae_doctrinae.scraping import monergism_scrap_general_information
         
         
 
@@ -36,6 +38,21 @@ def test_scrap_author_main_information():
     )
     ob.scrap_and_write()
 
+
+def test_get_all_author_general_information():
+    root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
+    browse_by_type = "speaker"
+
+    ob = monergism_scrap_general_information.MonergismScrapGeneralInformation(
+        root_folder=root_folder,
+        browse_by_type=browse_by_type,
+    )
+
+    asyncio.run(ob.download(30))
+    ob.update_downloaded_and_to_download()
+    ob.write_log_file()
+
+
 def test_get_author_all_work():
     root_folder ='D:/projet_github/FOR GOD/Scraping general/test_folder'
     browse_by_type = "speaker"
@@ -47,6 +64,9 @@ def test_get_author_all_work():
     ob.download(1)
     ob.update_downloaded_and_to_download()
     ob.write_log_file()
+
+
+
     
 
 
@@ -73,4 +93,5 @@ if __name__ == "__main__":
     # )
     # ob.scrap_and_write()
     #test_scrap_author_main_information()
-    test_get_author_all_work()
+    #test_get_author_all_work()
+    test_get_all_author_general_information()
